rename and add tests
