+++ tests/fixtures/js-stubs/teamIssuesBreakdown.js
@@ -1,0 +1,258 @@
+export function TeamIssuesBreakdown() {
+  return {
+    2: {
+      '2021-11-19T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-11-20T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-11-21T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-11-22T00:00:00+00:00': {
+        deleted: 11,
+        ignored: 11,
+        resolved: 11,
+        unignored: 11,
+        regressed: 11,
+        new: 11,
+        total: 11,
+      },
+      '2021-11-23T00:00:00+00:00': {
+        deleted: 7,
+        ignored: 7,
+        resolved: 7,
+        unignored: 7,
+        regressed: 7,
+        new: 7,
+        total: 7,
+      },
+      '2021-11-24T00:00:00+00:00': {
+        deleted: 4,
+        ignored: 4,
+        resolved: 4,
+        unignored: 4,
+        regressed: 4,
+        new: 4,
+        total: 4,
+      },
+      '2021-11-25T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-11-26T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-11-27T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-11-28T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-11-29T00:00:00+00:00': {
+        deleted: 8,
+        ignored: 8,
+        resolved: 8,
+        unignored: 8,
+        regressed: 8,
+        new: 8,
+        total: 8,
+      },
+      '2021-11-30T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-01T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-02T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-03T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-04T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-05T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-06T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-07T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-08T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-09T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-10T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-11T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-12T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-13T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-14T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-15T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+      '2021-12-16T00:00:00+00:00': {
+        deleted: 0,
+        ignored: 0,
+        resolved: 0,
+        unignored: 0,
+        regressed: 0,
+        new: 0,
+        total: 0,
+      },
+    },
+  };
+}
