--- tests/fixtures/js-stubs/teamIssuesReviewed.js
@@ -1,62 +1,0 @@
-export function TeamIssuesReviewed() {
-  return {
-    2: {
-      '2021-08-11T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-12T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-13T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-14T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-15T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-16T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-17T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-18T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-19T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-20T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-21T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-22T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-23T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-24T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-25T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-26T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-27T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-28T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-29T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-30T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-08-31T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-01T00:00:00+00:00': {reviewed: 10, total: 20},
-      '2021-09-02T00:00:00+00:00': {reviewed: 0, total: 10},
-      '2021-09-03T00:00:00+00:00': {reviewed: 1, total: 10},
-      '2021-09-04T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-05T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-06T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-07T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-08T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-09T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-10T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-11T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-12T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-13T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-14T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-15T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-16T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-17T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-18T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-19T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-20T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-21T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-22T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-23T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-24T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-25T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-26T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-27T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-28T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-29T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-09-30T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-10-01T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-10-02T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-10-03T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-10-04T00:00:00+00:00': {reviewed: 0, total: 0},
-      '2021-10-05T00:00:00+00:00': {reviewed: 0, total: 0},
-    },
-  };
-}
